ito-a-dashboard/Amit/ITOA_Dashboard/src/backend/requirements.txt
fastapi
pandas
uvicorn
textblob
wordcloud
scikit-learn
python-multipart
passlib[bcrypt]
python-jose[cryptography]
authlib
httpx
python-dotenv
itsdangerous
google-generativeai
requests